Security checks and REST processes integration

+++ backend/app/Http/Controllers/Api/EditorController.php
@@ -2,72 +2,54 @@
 
 namespace App\Http\Controllers\api;
 
+use App\Http\Requests\EditorConferenceYearUpdateRequest;
+use App\Http\Requests\EditorRequest;
 use App\Models\ConferenceYear;
 use App\Models\User;
 use App\Models\UserConferenceYear;
-use Illuminate\Http\Request;
+use App\Services\UserService;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 
 class EditorController extends Controller
 {
+    protected UserService $userService;
+
+    public function __construct(UserService $userService)
+    {
+        $this->userService = $userService;
+    }
     public function index() {
       $editors = User::where('role', 'editor')->with('conferenceYears')->get();
       return response()->json($editors);
     }
 
-    public function store(Request $request)
+    public function store(EditorRequest $request)
     {
-        $request->validate([
-        'email' => 'required|email|unique:users,email',
-        'name' => 'required|string|max:255|min:2',
-        'password' => 'required|string|min:6',
-        'conference_year_id' => 'required|exists:conference_years,id',
-        ]);
-
-        $editor = User::create([
-        'email' => $request->email,
-        'name' => $request->name,
-        'role' => 'editor',
-        'password' => bcrypt($request->password),
-        ]);
-
-        $editor->conferenceYears()->attach($request->conference_year_id);
-
-        return response()->json(['message' => 'Editor created', 'data' => $editor]);
+        Gate::authorize('create', User::class);
+        $data = $this->userService->createEditor($request);
+        return response()->json(['message' => $data['message'], 'data' => $data['object']]);
     }
 
-    public function update(Request $request, User $editor)
+    public function update(EditorConferenceYearUpdateRequest $request, User $editor)
     {
-        $request->validate([
-            'conferenceYearId' => 'required|exists:conference_years,id',
-        ]);
-
-        $editor->conferenceYears()->sync([$request->conferenceYearId]);
-
-        return response()->json(['message' => 'Editor updated', 'data' => $editor->load('conferenceYears')]);
+        Gate::authorize('update', $editor);
+        $data = $this->userService->updateEditorConferenceYear($request, $editor);
+        return response()->json(['message' => $data['message'], 'data' => $data['object']]);
     }
 
     public function destroy(User $editor)
     {
+        Gate::authorize('delete', $editor);
         $editor->delete();
         return response()->json(['message' => 'Editor deleted', 'data' => $editor]);
     }
 
     public function getYear() {
-        $user = auth('api')->user();
-
-        if (!$user) {
-            return response()->json(['error' => 'Unauthorized'], 401);
+        $data = $this->userService->getEditorYear();
+        if (!$data['success']) {
+            return response()->json(['message' => $data['message']], 401);
         }
-
-        if ($user->role == "admin") {
-            return response()->json(['message' => 'All']);
-        } else if($user->role == "editor") {
-            $yearId = UserConferenceYear::where("user_id", $user->id)->value("conference_year_id");
-            $year = ConferenceYear::find($yearId);
-            return response()->json($year);
-        } else {
-            return response()->json(['error' => 'Unauthorized'], 401);
-        }
+        return response()->json(['message' => $data['message']]);
     }
 }
